Crate lumol_input

This crate provide a way to build a Lumol simulation using input files.

Instead of building the System object, the Simulation object by hand before being able to use them to run the simulation, this crate allow to describe the simulation and the system in a TOML input file, using a simple syntax.

The main entry point of this crate is the Input struct, which allow to read a whole simulation configuration. The easiest way to run a simulation using Lumol is:

extern crate lumol_input;
use lumol_input::Input;

fn main() {
    let input = Input::new("simulation.toml").unwrap();
    let mut config = input.read().unwrap();

    config.simulation.run(&mut config.system, config.nsteps);
}

This crate also provide an InteractionsInput for reading interactions from a TOML file. It can be used to set the interactions in a system:

extern crate lumol;
extern crate lumol_input;
use lumol::System;
use lumol_input::InteractionsInput;

fn main() {
    let mut system = System::new();

    // ... Build the system by hand

    // Read the interactions
    let input = InteractionsInput::new("potentials.toml").unwrap();
    input.read(&mut system).unwrap();
}

Structs

Input
An input file for Lumol.

InteractionsInput
An interaction input file for Lumol.

Enums

Error
Possible causes of error when reading input files

Traits

FromToml
Convert a TOML table to a Rust type.

FromTomlWithData
Convert a TOML table and some additional data to a Rust type.

Type Definitions

Result
Custom Result type for input files