Struct lumol_input::Config

pub struct Config {
    pub system: System,
    pub simulation: Simulation,
    pub nsteps: usize,
}

A configuration about how to run a single simulation. This contains the system to simulate, the simulation itself and the number of steps to run the simulation.

Fields
system: System

The simulated system

simulation: Simulation

The simulation object

nsteps: usize

The simulation duration