Struct lumol_input::Input

pub struct Input { /* fields omitted */ }

An input file for Lumol.

Methods

impl Input

fn read_system(&self) -> Result<System>

Get the the simulated system. This is an internal function, public because of the code organization.

impl Input

fn read_outputs(&self) -> Result<Vec<(Box<Output>, u64)>>

Get the the simulation outputs. This is an internal function, public because of the code organization.

impl Input

fn read_propagator(&self) -> Result<Box<Propagator>>

Get the the simulation propagator. This is an internal function, public because of the code organization.

impl Input

fn read_simulation(&self) -> Result<Simulation>

Get the the simulation. This is an internal function, public because of the code organization.

fn read_nsteps(&self) -> Result<usize>

Get the number of steps in the simulation. This is an internal function, public because of the code organization.

fn simulation_table(&self) -> Result<&Table>

Get the simulation TOML table. This is an internal function, public because of the code organization.

impl Input

fn new<P: Into<PathBuf>>(path: P) -> Result<Input>

Read the file at Path and create a new Input from it.

fn read(&self) -> Result<Config>

Read input file and get the corresponding Config