Struct lumol_input::InteractionsInput

pub struct InteractionsInput { /* fields omitted */ }

An interaction input file for Lumol.

Methods

impl InteractionsInput

fn read_pairs(&self, system: &mut System) -> Result<()>

Read the "pairs" section from the potential configuration. This is an internal function, public because of the code organization.

fn read_bonds(&self, system: &mut System) -> Result<()>

Read the "bonds" section from the potential configuration. This is an internal function, public because of the code organization.

impl InteractionsInput

fn read_angles(&self, system: &mut System) -> Result<()>

Read the "angles" section from the potential configuration. This is an internal function, public because of the code organization.

fn read_dihedrals(&self, system: &mut System) -> Result<()>

Read the "dihedrals" section from the potential configuration. This is an internal function, public because of the code organization.

impl InteractionsInput

fn read_coulomb(&self, system: &mut System) -> Result<()>

Read the "coulomb" section from the potential configuration. This is an internal function, public because of the code organization.

fn read_charges(&self, system: &mut System) -> Result<()>

Read the "charges" from the potential configuration. This is an internal function, public because of the code organization.

impl InteractionsInput

fn new<P: Into<PathBuf>>(path: P) -> Result<InteractionsInput>

Read interactions from the TOML formatted file at path.

fn from_string(string: &str) -> Result<InteractionsInput>

Read the interactions from a TOML formatted string.

fn from_toml(config: Table) -> Result<InteractionsInput>

Read the interactions from a TOML table. This is an internal function, public because of the code organization.

fn read(&self, system: &mut System) -> Result<()>

Read the interactions from this input into the system.