Trait lumol_input::FromTomlWithData

pub trait FromTomlWithData: Sized {
    type Data;
    fn from_toml(table: &Table, data: Self::Data) -> Result<Self>;
}

Convert a TOML table and some additional data to a Rust type.

Associated Types

type Data

The type of the additional data needed.

Required Methods

fn from_toml(table: &Table, data: Self::Data) -> Result<Self>

Do the conversion from table and data to Self.